+++ flexmeasures/api/v3_0/accounts.py
@@ -36,7 +36,7 @@ class AccountAPI(FlaskView):
         .. :quickref: Account; Download account list
 
         This endpoint returns all accessible accounts.
-        Accessible accounts are your own account, or all accounts for admins.
+        Accessible accounts are your own account, the accounts you are a consultant for, or all accounts for admins.
         When the super-account concept (GH#203) lands, then users in such accounts see all managed accounts.
 
         **Example response**
@@ -50,6 +50,7 @@ class AccountAPI(FlaskView):
                     'id': 1,
                     'name': 'Test Account'
                     'account_roles': [1, 3],
+                    'consultancy_account_id': 2,
                 }
             ]
 
@@ -65,7 +66,12 @@ class AccountAPI(FlaskView):
         if user_has_admin_access(current_user, "read"):
             accounts = get_accounts()
         else:
-            accounts = [current_user.account]
+            accounts = [current_user.account] + (
+                current_user.account.consultant_client_accounts
+                if "customer-manager" in current_user.roles
+                else []
+            )
+
         return accounts_schema.dump(accounts), 200
 
     @route("/<id>", methods=["GET"])
@@ -78,7 +84,7 @@ class AccountAPI(FlaskView):
         .. :quickref: Account; Get an account
 
         This endpoint retrieves an account, given its id.
-        Only admins or the user themselves can use this endpoint.
+        Only admins, consultant users or the user themselves can use this endpoint.
 
         **Example response**
 
@@ -88,6 +94,8 @@ class AccountAPI(FlaskView):
                 'id': 1,
                 'name': 'Test Account'
                 'account_roles': [1, 3],
+                'consultancy_account_id':2,
+                'consultant_name':'Consultant',
             }
 
         :reqheader Authorization: The authentication token
@@ -99,4 +107,5 @@ class AccountAPI(FlaskView):
         :status 403: INVALID_SENDER
         :status 422: UNPROCESSABLE_ENTITY
         """
+
         return account_schema.dump(account), 200