+++ documentation/changelog.rst
@@ -14,6 +14,7 @@ New features
 
 - Different site-level production and consumption limits can be defined for the storage scheduler via the API (``flex-context``) or via asset attributes [see `PR #884 <https://github.com/FlexMeasures/flexmeasures/pull/884>`_]
 - Scheduling data better distinguishes (e.g. in chart tooltips) when a schedule was the result of a fallback mechanism, by splitting off the fallback mechanism from the main scheduler (as a separate job) [see `PR #846 <https://github.com/FlexMeasures/flexmeasures/pull/846>`_]
+- A Consultancy account can be added to any account, with the effect that the customer-manager users in that account get read access. [see `PR #846 <https://github.com/FlexMeasures/flexmeasures/pull/877>`_]
 
 Infrastructure / Support
 ----------------------